Connection error and general exception with socket error management and await timeout

+++ src/main/kotlin/com/raycoarana/memkched/internal/SocketConfig.kt
@@ -3,7 +3,7 @@ package com.raycoarana.memkched.internal
 internal data class SocketConfig(
     val inBufferSize: Int,
     val outBufferSize: Int,
-    val readTimeout: Int,
-    val writeTimeout: Int,
+    val readTimeout: Long,
+    val writeTimeout: Long,
     val nioThreadPoolInitialSize: Int
 )
